Info moved to special directory, writed GenerateReport method
